% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractSec.R
\name{fractSec}
\alias{fractSec}
\title{Extract Fractional Seconds from Time Stamps}
\usage{
fractSec(time.stamps)
}
\arguments{
\item{time.stamps}{A vector of detection time stamps in \code{POSIXct} format.}
}
\value{
A numeric vector of fractional seconds (0 to 1 range).
}
\description{
This function extracts the fractional second component from timestamps in
\code{POSIXct} format.
}
\examples{
time_stamps <- as.POSIXct(c("2025-02-26 12:00:00.123",
                            "2025-02-26 12:00:01.456"))
fs <- fractSec(time_stamps)
print(fs)

}
