% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwrapDrift.R
\name{unwrapDrift}
\alias{unwrapDrift}
\title{Unwrap Time Drift in Acoustic Telemetry Data}
\usage{
unwrapDrift(fs, time.stamps, granularity = 1)
}
\arguments{
\item{fs}{A numeric vector of fractional seconds (0-1 range).}

\item{time.stamps}{A \code{POSIXct} vector of detection time stamps.}

\item{granularity}{A numeric value specifying the emission granularity of the
transmitter. Defaults to \code{1} for Thelma Biotel transmitters. Use \code{0.5} for
Lotek transmitters.}
}
\value{
A numeric vector representing the unwrapped, continuous time drift
values (in seconds), starting at 0.
}
\description{
This function reconstructs continuous drift values from a vector of
fractional seconds, removing discontinuities caused by the 0–1 wrapping
of fractional second data and by emission granularities smaller than 1
second.
}
\details{
The unwrapping process starts by estimating a general linear drift rate
(slope) from the detection timestamps. This slope is used to predict the
expected position of each detection along a continuous trend. Each observed
fractional second value is then adjusted—by shifting it by a multiple of
the transmitter's emission granularity—so that it aligns with the predicted
continuous trend.
}
\examples{
# Load detection data and subset one individual and deployment
data(fish_data)
detect <- fish_data$detection_data
detect <- detect[fish_id == "CORHIP-892" & deployment_id == 42]

# Extract fractional seconds (0-1 constrained)
detect$fs <- fractSec(time.stamps = detect$date_time)
plot(fs ~ date_time, data = detect)

# Unwrap constrained drift (granularity = 1)
detect$drift <- unwrapDrift(fs = detect$fs, time.stamps = detect$date_time,
                            granularity = 1)
plot(drift ~ date_time, data = detect)

}
